NEXT-7703 Create CookieService to add cookie groups to storefront
